final IL-6 + modified packages names
