+++ backend/src/main/java/com/agh/introwertycznelosie/IntrowertyczneLosieApplication.java
@@ -1,4 +1,4 @@
-package com.agh.introwertycznelosie;
+package main.java.com.agh.introwertycznelosie;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
