+++ backend/src/main/java/com/agh/introwertycznelosie/data/Major.java
@@ -1,0 +1,107 @@
+package main.java.com.agh.introwertycznelosie.data;
+
+import main.java.com.agh.introwertycznelosie.data.Faculty;
+import main.java.com.agh.introwertycznelosie.data.ModeOfStudy;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+public class Major {
+
+    @Id
+    @GeneratedValue
+    private int id;
+
+    private Faculty faculty;
+    private String fullName;
+    private String shortName;
+    private ModeOfStudy mode;
+    private int numberOfPlaces;
+
+    // TODO - klasa zamiast stringa, ale klasy jeszcze nie ma
+    private String contactPerson1;
+    private String contactPerson2;
+
+    private boolean mixedField;
+    private String annotations;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public Faculty getFaculty() {
+        return faculty;
+    }
+
+    public void setFaculty(Faculty faculty) {
+        this.faculty = faculty;
+    }
+
+    public String getFullName() {
+        return fullName;
+    }
+
+    public void setFullName(String fullName) {
+        this.fullName = fullName;
+    }
+
+    public String getShortName() {
+        return shortName;
+    }
+
+    public void setShortName(String shortName) {
+        this.shortName = shortName;
+    }
+
+    public ModeOfStudy getMode() {
+        return mode;
+    }
+
+    public void setMode(ModeOfStudy mode) {
+        this.mode = mode;
+    }
+
+    public int getNumberOfPlaces() {
+        return numberOfPlaces;
+    }
+
+    public void setNumberOfPlaces(int numberOfPlaces) {
+        this.numberOfPlaces = numberOfPlaces;
+    }
+
+    public String getContactPerson1() {
+        return contactPerson1;
+    }
+
+    public void setContactPerson1(String contactPerson1) {
+        this.contactPerson1 = contactPerson1;
+    }
+
+    public String getContactPerson2() {
+        return contactPerson2;
+    }
+
+    public void setContactPerson2(String contactPerson2) {
+        this.contactPerson2 = contactPerson2;
+    }
+
+    public boolean isMixedField() {
+        return mixedField;
+    }
+
+    public void setMixedField(boolean mixedField) {
+        this.mixedField = mixedField;
+    }
+
+    public String getAnnotations() {
+        return annotations;
+    }
+
+    public void setAnnotations(String annotations) {
+        this.annotations = annotations;
+    }
+}
