+++ backend/src/main/java/com/agh/introwertycznelosie/data/ModeOfStudy.java
@@ -1,0 +1,6 @@
+package main.java.com.agh.introwertycznelosie.data;
+
+public enum ModeOfStudy {
+
+    fullTime, partTime;
+}
