+++ backend/src/main/java/com/agh/introwertycznelosie/data/Role.java
@@ -1,4 +1,4 @@
-package com.agh.introwertycznelosie.data;
+package main.java.com.agh.introwertycznelosie.data;
 
 import org.springframework.data.repository.cdi.Eager;
