+++ backend/src/main/java/com/agh/introwertycznelosie/data/User.java
@@ -1,4 +1,4 @@
-package com.agh.introwertycznelosie.data;
+package main.java.com.agh.introwertycznelosie.data;
 
 import javax.persistence.*;
 import java.util.Set;
