+++ backend/src/main/java/com/agh/introwertycznelosie/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
-package com.agh.introwertycznelosie.repositories;
+package main.java.com.agh.introwertycznelosie.repositories;
 
-import com.agh.introwertycznelosie.data.Role;
+import main.java.com.agh.introwertycznelosie.data.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long> { }
