+++ backend/src/main/java/com/agh/introwertycznelosie/services/UserDetailsServiceImpl.java
@@ -1,13 +1,13 @@
-package com.agh.introwertycznelosie.services;
+package main.java.com.agh.introwertycznelosie.services;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-import com.agh.introwertycznelosie.data.Role;
-import com.agh.introwertycznelosie.data.User;
-import com.agh.introwertycznelosie.repositories.UserRepository;
+import main.java.com.agh.introwertycznelosie.data.Role;
+import main.java.com.agh.introwertycznelosie.data.User;
+import main.java.com.agh.introwertycznelosie.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
