+++ backend/src/test/java/com/agh/introwertycznelosie/IntrowertyczneLosieApplicationTests.java
@@ -1,4 +1,4 @@
-package com.agh.introwertycznelosie;
+package test.java.com.agh.introwertycznelosie;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
