--- src/main/java/com/agh/introwertycznelosie/data/ModeOfStudy.java
@@ -1,6 +1,0 @@
-package com.agh.introwertycznelosie.data;
-
-public enum ModeOfStudy {
-
-    fullTime, partTime;
-}
